requirements.txt
flask
Joking
requests
pillow